增加音频截取功能

+++ meaudiodecoder.h
@@ -37,7 +37,7 @@ public:
     QString getFileName();
     int getAudioIndex(){return audioindex;}
     int decoder(QVector<double>& retData);
-    int SeekFrame(int num=0);
+    int SeekFrame(int64_t num=0);
 private:
     char fileName[FILENAME_MAX];
     AVFormatContext *infmt_ctx;
